Restructure test endpoint to better match what the client would like to send
