+++ src/main/java/uk/ac/cam/cl/dtg/isaac/api/TestsFacade.java
@@ -1,14 +1,30 @@
+/*
+ * Copyright 2019 Meurig Thomas
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ *
+ * You may obtain a copy of the License at
+ * 		http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package uk.ac.cam.cl.dtg.isaac.api;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.api.client.util.Lists;
 import com.google.inject.Inject;
-import com.mongodb.util.JSONParseException;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.jboss.resteasy.annotations.GZIP;
+import org.jboss.resteasy.annotations.jaxrs.QueryParam;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import uk.ac.cam.cl.dtg.isaac.dto.TestCaseDTO;
 import uk.ac.cam.cl.dtg.isaac.dto.TestDTO;
 import uk.ac.cam.cl.dtg.segue.api.managers.QuestionManager;
 import uk.ac.cam.cl.dtg.segue.api.managers.UserAccountManager;
@@ -30,13 +46,13 @@ import javax.ws.rs.BadRequestException;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.POST;
 import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import java.io.IOException;
 import java.util.List;
+
 
 // TODO documentation
 @Path("/tests")
@@ -56,14 +72,14 @@ public class TestsFacade extends AbstractIsaacFacade {
     }
 
     @POST
-    @Path("/{questionType}")
+    @Path("/question")
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
     @GZIP
     @ApiOperation(value = "Create a new user or update an existing user.")
     public Response createOrUpdateUserSettings(@Context final HttpServletRequest request,
                                                @Context final HttpServletResponse response,
-                                               @PathParam("questionType") final String questionType,
+                                               @QueryParam("type") final String questionType,
                                                final String testJson) {
         try {
             // User authorisation
@@ -77,7 +93,7 @@ public class TestsFacade extends AbstractIsaacFacade {
 
             // Create a fake question
             Class<? extends Content> questionClass = mapper.getClassByType(questionType);
-            if (null == questionClass || ChoiceQuestion.class.isAssignableFrom(questionClass)) {
+            if (null == questionClass || !ChoiceQuestion.class.isAssignableFrom(questionClass)) {
                 throw new BadRequestException(String.format("Not a valid questionType (%s)", questionType));
             }
             ChoiceQuestion testQuestion = (ChoiceQuestion) questionClass.newInstance();
@@ -89,8 +105,8 @@ public class TestsFacade extends AbstractIsaacFacade {
             }
 
             // For each test, check its results against the fake question
-            List<TestDTO.TestCaseDTO> results = Lists.newArrayList();
-            for (TestDTO.TestCaseDTO testCase : test.getCases()) {
+            List<TestCaseDTO> results = Lists.newArrayList();
+            for (TestCaseDTO testCase : test.getTestCases()) {
                 QuestionValidationResponse questionValidationResponse =
                         questionValidator.validateQuestionResponse(testQuestion, testCase.getChoice());
                 testCase.setActual(questionValidationResponse.isCorrect());
@@ -104,7 +120,7 @@ public class TestsFacade extends AbstractIsaacFacade {
             return SegueErrorResponse.getNotLoggedInResponse();
         } catch (ValidatorUnavailableException | IOException e) {
             return SegueErrorResponse.getServiceUnavailableResponse(e.getMessage());
-        } catch (JSONParseException | InstantiationException | IllegalAccessException e) {
+        } catch (InstantiationException | IllegalAccessException e) {
             return SegueErrorResponse.getBadRequestResponse(e.getMessage());
         }
     }
