+++ src/main/java/uk/ac/cam/cl/dtg/isaac/dto/TestDTO.java
@@ -1,4 +1,4 @@
-/**
+/*
  * Copyright 2019 Meurig Thomas
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -17,61 +17,27 @@ package uk.ac.cam.cl.dtg.isaac.dto;
 
 
 import uk.ac.cam.cl.dtg.segue.dos.content.Choice;
-import uk.ac.cam.cl.dtg.segue.dos.content.Content;
 
 import java.util.List;
 
+/**
+ * DTO for constructing a test question and the answer cases we expect
+ */
 public class TestDTO {
-    /**
-     * DTO for holding choices and their expected values
-     */
-    public static final class TestCaseDTO {
-        Choice choice;
-        Boolean expected;
-        Boolean actual;
-        Content explanation;
-
-        public void setChoice(Choice choice) {
-            this.choice = choice;
-        }
-        public Choice getChoice() {
-            return this.choice;
-        }
-
-        public void setExpected(Boolean expected) {
-            this.expected = expected;
-        }
-        public Boolean getExpected() {
-            return this.expected;
-        }
-
-        public void setActual(Boolean actual) {
-            this.actual = actual;
-        }
-        public Boolean getActual() {
-            return this.actual;
-        }
-
-        public void setExplanation(Content explanation) {
-            this.explanation = explanation;
-        }
-        public Content getExplanation() {
-            return this.explanation;
-        }
-    }
-
     List<Choice> choices;
-    List<TestCaseDTO> tests;
+    List<TestCaseDTO> testCases;
+
     public void setChoices(List<Choice> choices) {
         this.choices = choices;
     }
     public List<Choice> getChoices() {
         return this.choices;
     }
-    public void setCases(List<TestCaseDTO> tests) {
-        this.tests = tests;
+
+    public void setTestCases(List<TestCaseDTO> tests) {
+        this.testCases = tests;
     }
-    public List<TestCaseDTO> getCases() {
-        return this.tests;
+    public List<TestCaseDTO> getTestCases() {
+        return this.testCases;
     }
 }
